feat: remove marshmallow type
